duplicate controller and service
